Add files via upload
The project is about finding 3 phase circuit parameters. In a nutshell it is a project or can be used as an app for finding the desired parameters. The goal in this project is to create a three-phase circuit analyzer using MATLAB and GUI. We’ve created an analyzer app that takes inputs from the user in a graphic interface and returns the values of line currents, phase currents, line voltages, phase voltages, total power, real power, reactive power and power factor of the loads. This circuit analyzer works for both balanced and unbalanced three
phase circuits.